Rename `workflow.py` -> `core.py`.
