+++ gwf/__init__.py
@@ -1,5 +1,5 @@
 from __future__ import absolute_import, print_function
 
-from gwf.workflow import Target, Workflow
+from gwf.core import Target, Workflow
 
 __all__ = ('Target', 'Workflow')
